+++ examples/fetch_and_cat.rs
@@ -39,7 +39,7 @@ async fn main() -> anyhow::Result<()> {
 
     // Initialize the repo and start a daemon.
     // UninitializedIpfs will handle starting up the repository and return the facade (ipfs::Ipfs)
-    let ipfs: Ipfs = UninitializedIpfs::new().start().await?;
+    let ipfs: Ipfs = UninitializedIpfs::new().with_default().start().await?;
 
     if opt.default_bootstrappers {
         // applications wishing to find content on the global IPFS swarm should restore the latest