+++ examples/ipns-publish.rs
@@ -7,8 +7,9 @@ async fn main() -> anyhow::Result<()> {
 
     // Initialize the repo and start a daemon
     let ipfs: Ipfs = UninitializedIpfs::new()
-        .enable_mdns()
-        .enable_upnp()
+        .with_default()
+        .with_mdns()
+        .with_relay(true)
         .default_record_key_validator()
         .start()
         .await?;