+++ examples/ipns-resolve.rs
@@ -19,7 +19,9 @@ async fn main() -> anyhow::Result<()> {
 
     // Initialize the repo and start a daemon
     let ipfs: Ipfs = UninitializedIpfs::new()
-        .enable_mdns()
+        .with_default()
+        .with_mdns()
+        .with_relay(true)
         .default_record_key_validator()
         .start()
         .await?;