+++ src/p2p/mod.rs
@@ -1,18 +1,14 @@
 //! P2P handling for IPFS nodes.
 use std::convert::TryInto;
 use std::num::{NonZeroU8, NonZeroUsize};
-use std::time::Duration;
 
 use crate::error::Error;
 use crate::repo::Repo;
 use crate::{IpfsOptions, TTransportFn};
 
-use either::Either;
 use libp2p::gossipsub::ValidationMode;
 use libp2p::identify::Info as IdentifyInfo;
 use libp2p::identity::{Keypair, PublicKey};
-use libp2p::kad::KademliaConfig;
-use libp2p::ping::Config as PingConfig;
 use libp2p::swarm::NetworkBehaviour;
 use libp2p::{Multiaddr, PeerId};
 use libp2p::{StreamProtocol, Swarm};
@@ -107,99 +103,6 @@ impl From<IdentifyInfo> for PeerInfo {
     }
 }
 
-/// Defines the configuration for an IPFS swarm.
-pub struct SwarmOptions {
-    /// The peers to connect to on startup.
-    pub bootstrap: Vec<Multiaddr>,
-    /// Enables mdns for peer discovery and announcement when true.
-    pub mdns: bool,
-    /// enables ipv6 for mdns
-    pub mdns_ipv6: bool,
-    /// disable kad
-    pub disable_kad: bool,
-    /// disable bitswap
-    pub disable_bitswap: bool,
-    /// Relay Server
-    pub relay_server: bool,
-    /// Relay Server Configuration
-    pub relay_server_config: Option<RelayConfig>,
-    /// Kademlia Configuration
-    pub kad_config: Option<Either<KadConfig, KademliaConfig>>,
-    /// Ping Configuration
-    pub ping_config: Option<PingConfig>,
-    /// bitswap config
-    pub bitswap_config: Option<BitswapConfig>,
-    /// identify configuration
-    pub identify_config: Option<IdentifyConfiguration>,
-    /// Kad store config
-    /// Note: Only supports MemoryStoreConfig at this time
-    pub kad_store_config: KadStoreConfig,
-    /// Pubsub configuration,
-    pub pubsub_config: Option<PubsubConfig>,
-    /// addressbook config
-    pub addrbook_config: Option<AddressBookConfig>,
-    /// UPnP/PortMapping
-    pub portmapping: bool,
-    /// Relay client
-    pub relay: bool,
-    /// Enables dcutr
-    pub dcutr: bool,
-    /// Connection idle
-    pub connection_idle: Duration,
-    pub rendezvous_client: bool,
-    pub rendezvous_server: bool,
-}
-
-impl From<&IpfsOptions> for SwarmOptions {
-    fn from(options: &IpfsOptions) -> Self {
-        let bootstrap = options.bootstrap.clone();
-        let mdns = options.mdns;
-        let mdns_ipv6 = options.mdns_ipv6;
-        let dcutr = options.dcutr;
-        let relay_server = options.relay_server;
-        let relay_server_config = options.relay_server_config.clone();
-        let relay = options.relay;
-        let kad_config = options.kad_configuration.clone();
-        let ping_config = options.ping_configuration.clone();
-        let kad_store_config = options.kad_store_config.clone();
-        let disable_kad = options.disable_kad;
-        let disable_bitswap = options.disable_bitswap;
-        let bitswap_config = options.bitswap_config.clone();
-
-        let identify_config = options.identify_configuration.clone();
-        let portmapping = options.port_mapping;
-        let pubsub_config = options.pubsub_config.clone();
-        let addrbook_config = options.addr_config;
-
-        let connection_idle = options.connection_idle;
-        let rendezvous_client = options.rendezvous_client;
-        let rendezvous_server = options.rendezvous_server;
-
-        SwarmOptions {
-            bootstrap,
-            mdns,
-            disable_kad,
-            disable_bitswap,
-            bitswap_config,
-            mdns_ipv6,
-            relay_server,
-            relay_server_config,
-            relay,
-            dcutr,
-            kad_config,
-            kad_store_config,
-            ping_config,
-            identify_config,
-            portmapping,
-            addrbook_config,
-            pubsub_config,
-            connection_idle,
-            rendezvous_client,
-            rendezvous_server,
-        }
-    }
-}
-
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct PubsubConfig {
     /// Custom protocol name
@@ -277,7 +180,7 @@ impl Default for SwarmConfig {
 /// Creates a new IPFS swarm.
 pub async fn create_swarm<C>(
     keypair: &Keypair,
-    options: SwarmOptions,
+    options: &IpfsOptions,
     swarm_config: SwarmConfig,
     transport_config: TransportConfig,
     repo: Repo,
@@ -294,8 +197,7 @@ where
     let idle = options.connection_idle;
 
     let (behaviour, relay_transport) =
-        behaviour::build_behaviour(&keypair, options, repo, swarm_config.connection, custom)
-            .await?;
+        behaviour::Behaviour::new(&keypair, options, repo, swarm_config.connection, custom).await?;
 
     // Set up an encrypted TCP transport over the Yamux and Mplex protocol. If relay transport is supplied, that will be apart
     let transport = match custom_transport {